pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.4.0",
    "discord-webhook>=1.3.1",
    "flask>=3.1.0",
    "requests>=2.32.3",
]
